requirements.txt
lancedb 
tantivy 
pypdf 
sqlalchemy
phidata
streamlit
sentence-transformers
langchain
langchain-community
smolagents
litellm
datasets
chromadb
faiss-cpu
#faiss-gpu
pypdf
langchain_huggingface
langchain_aws
transformers>=4.48.0
pytest
boto3
praisonaiagents[knowledge]
ollama
qdrant-client
langchain_ollama
pdfplumber
langchain-google-vertexai
llama-index
llama-index-llms-groq
groq
rank_bm25
cohere
langchain-cohere
flask
langchain-unstructured
PyMuPDF
pytesseract
Pillow
tesseract